+++ src/pages/Cdrs/Cdrs.vue
@@ -14,7 +14,7 @@ import { isEmpty, flow, get } from 'lodash';
 import utils from '@/utils';
 import { CDRS } from '@/constants';
 
-import DataTable from '../DataTable/DataTable';
+import DataTable from '@/components/DataTable/DataTable';
 import { TABLE_HEADERS } from './constants';
 
 import 'vue2-daterange-picker/dist/vue2-daterange-picker.css';
