+++ src/pages/Rates/constants.js
