+++ src/pages/StatisticsCharts/fixtures.js
